big commit 3

+++ src/expression/exceptions/SignedMoveR.java
@@ -1,7 +1,6 @@
 package expression.exceptions;
 
-import java.math.BigDecimal;
-import expression.*;
+import expression.ExpressionPart;
 
 public class SignedMoveR extends AbstractCheckedOperation implements ExpressionPart {
     public SignedMoveR(ExpressionPart p1, ExpressionPart p2) {
@@ -16,7 +15,7 @@ public class SignedMoveR extends AbstractCheckedOperation implements ExpressionP
         return ">>>";
     }
 
-    public long calc(int num1, int num2) {
+    public int calc(int num1, int num2) {
         return num1 >>> num2;
     }
